New OpenBot
Rigged with a Rigify Rig, but optimized for Game Engines for easy import (via the EXPY blender plugin). Should be the best of both worlds. Using the MannyQuin-BoneMap.res file will allow this skeleton to easily load animation libraries in Godot. NOTE, although these bones are the same as Rigify's bones, the structure of Rigify's bones make it severely deform without EXPY's Game Engine optimizer. That optimization has already been applied to this file so the plugin isn't entirely necessary for this file or for making new Godot animations with it.